pyproject.toml
[tool.poetry]
name = "lucid_test"
version = "0.1.0"
description = ""
authors = ["Vadym Markov"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
fastapi-events = "^0.10.1"
sqlmodel = "^0.0.14"
uvicorn = "^0.27.1"
PyJWT = "^2.8.0"
pymysql = "^1.1.1"
cryptography = "^44.0.2"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
mypy = "^1.8.0"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
